+++ app/renderer/app_status.js
@@ -2,7 +2,6 @@
 
 const EventEmitter = require('events');
 
-const TasksHandler = require('./tasks_handler');
 const version = require('../../package.json').version;
 const AppConfigStatus = require('./app_config_status');
 
@@ -10,10 +9,6 @@ module.exports = {
     editMode: false,
     activeSuite: 0,
     events: new EventEmitter(),
-    toggleEdit() {
-        this.editMode = !this.editMode;
-        TasksHandler.saveTasks();
-    },
     config: new AppConfigStatus(),
     maxSuites: 6,
     maxTasksPerSuite: 8,