+++ app/renderer/components/navbar.js
@@ -1,12 +1,13 @@
 "use strict";
 
 const Suite = require('../../common/suite');
-const Material = require('../materialize');
+const Material = require('../api/materialize');
 const AppStatus = require('../app_status');
+const GauchoActions = require('../api/gaucho_actions');
 const NavbarMenu = require('./navbar_menu');
 const TapTarget = require('./tap_target');
 const ContextMenu = require('./context_menu');
-const DeleteConfirmationAlert = require('../app_alerts').DeleteConfirmationAlert;
+const DeleteConfirmationAlert = require('../api/app_alerts').DeleteConfirmationAlert;
 
 module.exports = {
     props: ['suites'],
@@ -14,11 +15,6 @@ module.exports = {
         "navbar-menu": NavbarMenu,
         "tap-target": TapTarget,
         "context-menu": ContextMenu
-    },
-    data() {
-        return {
-            AppStatus: AppStatus
-        };
     },
     template: `
     <div>
@@ -88,7 +84,7 @@ module.exports = {
             });
         },
         toggleEdit() {
-            AppStatus.toggleEdit();
+            GauchoActions.toggleEdit();
         },
         openAbout() {
             this.$refs.navMenu.$refs.about.openAbout();