+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   * Tasks can now be reordered within a suite while in edit mode
   * Added an initial message, noting the edit button to new users
   * Tasks with no path defined will now be executed on user home directory
-  * Added "export tasks" button in config
+  * Added import and export options in config
   * Added confirmation popup when deleting a task, suite or clearing all tasks
   * Add Task form now collapses after adding a task
   * About menu style improved
@@ -13,10 +13,12 @@
   * Fixed bug that prevented running tasks to be killed when closing Gaucho
   * Fixed bug where deleting a suite doesn't update total tasks in the bottom bar
   * Fixed bug where clearing all tasks doesn't update total tasks
+  * Fixed bug where clearing all tasks sometimes broke the Add-task button
   * Added pre-commit hook
   * Added a dev config environment
   * Code refactoring in several components and modules
   * Code linting rules changed, using eslint instead of jshint
+  * Wrote more descriptive and appropriate alert messages
 
 0.3.0 / 2017-08-29
 ==================